+++ lib/screens/record_screen/ranking_tab.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:untitled/utils/size_config.dart';
+import 'package:untitled/config/size_config.dart';
 
 import '../../utils/tmp_mock_data.dart';
